+++ system/modules/core/models/FilesModel.php
@@ -176,6 +176,18 @@ class FilesModel extends \Model
 			$strUuid = \StringUtil::uuidToBin($strUuid);
 		}
 
+		// Check the model registry (does not work by default due to UNHEX())
+		if (empty($arrOptions))
+		{
+			/** @var \FilesModel $objModel */
+			$objModel = \Model\Registry::getInstance()->fetch(static::$strTable, $strUuid, 'uuid');
+
+			if ($objModel !== null)
+			{
+				return $objModel;
+			}
+		}
+
 		return static::findOneBy(array("$t.uuid=UNHEX(?)"), bin2hex($strUuid), $arrOptions);
 	}